add error wrapper
