+++ packages/common/src/allo-strategy-adapters/adapter.ts
@@ -1,25 +1,5 @@
 import { Address, PublicClient } from "viem";
-
-export type AdapterNotFoundError = {
-  type: "ADAPTER_NOT_FOUND";
-  strategyName: string;
-  error: Error;
-};
-
-export type AdapterAllocationUnauthorizedError = {
-  type: "ADAPTER_ALLOCATION_UNAUTHORIZED";
-  error: Error;
-};
-
-export type AdapterUnknownError = {
-  type: "ADAPTER_UNKNOWN_ERROR";
-  error: Error;
-};
-
-export type AdapterErrorWrapper =
-  | AdapterNotFoundError
-  | AdapterAllocationUnauthorizedError
-  | AdapterUnknownError;
+import { AdapterErrorWrapper } from "./errorWrapper";
 
 export type AdapterResponse<T> =
   | { type: "success"; value: T }
